+++ packages/search-adapter/src/empathy/__tests__/request-mappers/empathy-request-filters-mapper.spec.ts
@@ -93,8 +93,7 @@ function getMockedFacet(name: string): Facet {
   return {
     modelName: name,
     title: name,
-    filters: [],
-    preselected: false
+    filters: []
   };
 }