+++ packages/search-adapter/src/empathy/models/requests/empathy-recommendations.request.ts
@@ -1,4 +1,4 @@
-export interface EmpathyRecommendationsRequest {
+export interface EmpathyTopRecommendationsRequest {
   catalogue?: string;
   lang: string;
   origin?: string;