+++ packages/search-adapter/src/types/search-adapter.types.ts
@@ -1,26 +1,40 @@
 import {
+  ClicksRecommendationsRequest,
   NextQueriesRequest,
-  RecommendationsRequest,
+  QueriesRecommendationsRequest,
   RelatedTagsRequest,
   RequestOptions,
   SearchByIdRequest,
   SearchRequest,
+  SectionRecommendationsRequest,
   SuggestionsRequest,
-  TrackingRequest
+  TopRecommendationsRequest,
+  TrackingRequest,
+  UserRecommendationsRequest
 } from './requests.types';
 import {
+  ClicksRecommendationsResponse,
   NextQueriesResponse,
-  RecommendationsResponse,
+  QueriesRecommendationsResponse,
   RelatedTagsResponse,
   SearchByIdResponse,
   SearchResponse,
-  SuggestionsResponse
+  SectionRecommendationsResponse,
+  SuggestionsResponse,
+  TopRecommendationsResponse,
+  UserRecommendationsResponse
 } from './response.types';
 
 export interface SearchAdapter {
   // Required functions
   getNextQueries(request: NextQueriesRequest, requestOptions?: RequestOptions): Promise<NextQueriesResponse>;
-  getRecommendations(request: RecommendationsRequest, requestOptions?: RequestOptions): Promise<RecommendationsResponse>;
+  getTopRecommendations(request: TopRecommendationsRequest, requestOptions?: RequestOptions): Promise<TopRecommendationsResponse>;
+  getSectionRecommendations(request: SectionRecommendationsRequest,
+    requestOptions?: RequestOptions): Promise<SectionRecommendationsResponse>;
+  getQueriesRecommendations(request: QueriesRecommendationsRequest,
+    requestOptions?: RequestOptions): Promise<QueriesRecommendationsResponse>;
+  getClicksRecommendations(request: ClicksRecommendationsRequest, requestOptions?: RequestOptions): Promise<ClicksRecommendationsResponse>;
+  getUserRecommendations(request: UserRecommendationsRequest, requestOptions?: RequestOptions): Promise<UserRecommendationsResponse>;
   getRelatedTags(request: RelatedTagsRequest, requestOptions?: RequestOptions): Promise<RelatedTagsResponse>;
   getSuggestions(request: SuggestionsRequest, requestOptions?: RequestOptions): Promise<SuggestionsResponse>;
   search(request: SearchRequest, requestOptions?: RequestOptions): Promise<SearchResponse>;